FastAPI/school/pyproject.toml
[tool.poetry]
name = "school"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
psycopg2-binary = "^2.9.5"
sqlalchemy = "^1.4.45"
asyncpg = "^0.27.0"
uvicorn = "^0.20.0"
pytz = "^2022.7"
passlib = "^1.7.4"
python-multipart = "^0.0.5"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}

[tool.pytest.ini_options]
pythonpath = [
  ".", "./course_api"
]

[tool.poetry.scripts]
school = 'manage:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
